+++ src/scene_test2.c
@@ -62,11 +62,11 @@ void scene_test2_update(SceneTest *scene, double delta) {
     scene->offset += 0.025;
     {
       int32_t offset_uniform = picasso_program_uniform_location(scene->layers[0]->program, "offset");
-      picasso_program_uniform_float(scene->layers[0]->program, offset_uniform, cos(scene->offset) / 2.0);
+      picasso_program_uniform_float(scene->layers[0]->program, offset_uniform, (float)(cos(scene->offset) / 2.0));
     }
     {
       int32_t offset_uniform = picasso_program_uniform_location(scene->layers[1]->program, "offset");
-      picasso_program_uniform_float(scene->layers[1]->program, offset_uniform, cos(scene->offset) / 2.0);
+      picasso_program_uniform_float(scene->layers[1]->program, offset_uniform, (float)(cos(scene->offset) / 2.0));
     }
   }
 }
